Removing pointless app utils
