--- src/shared/utils/app/calculate-upvote-pct.ts
@@ -1,6 +1,0 @@
-export default function calculateUpvotePct(
-  upvotes: number,
-  downvotes: number,
-): number {
-  return (upvotes / (upvotes + downvotes)) * 100;
-}
