--- src/shared/utils/app/color-list.ts
@@ -1,9 +1,0 @@
-export const colorList: string[] = [
-  "var(--comment-node-1-color)",
-  "var(--comment-node-2-color)",
-  "var(--comment-node-3-color)",
-  "var(--comment-node-4-color)",
-  "var(--comment-node-5-color)",
-  "var(--comment-node-6-color)",
-  "var(--comment-node-7-color)",
-];
