--- src/shared/utils/app/edit-comment-reply.ts
@@ -1,9 +1,0 @@
-import { editListImmutable } from "@utils/helpers";
-import { CommentReplyView } from "lemmy-js-client";
-
-export default function editCommentReply(
-  data: CommentReplyView,
-  replies: CommentReplyView[],
-): CommentReplyView[] {
-  return editListImmutable("comment_reply", data, replies);
-}
