--- src/shared/utils/app/edit-comment-report.ts
@@ -1,9 +1,0 @@
-import { editListImmutable } from "@utils/helpers";
-import { CommentReportView } from "lemmy-js-client";
-
-export default function editCommentReport(
-  data: CommentReportView,
-  reports: CommentReportView[],
-): CommentReportView[] {
-  return editListImmutable("comment_report", data, reports);
-}
