--- src/shared/utils/app/edit-comment.ts
@@ -1,9 +1,0 @@
-import { editListImmutable } from "@utils/helpers";
-import { CommentView } from "lemmy-js-client";
-
-export default function editComment(
-  data: CommentView,
-  comments: CommentView[],
-): CommentView[] {
-  return editListImmutable("comment", data, comments);
-}
