--- src/shared/utils/app/edit-post-report.ts
@@ -1,9 +1,0 @@
-import { editListImmutable } from "@utils/helpers";
-import { PostReportView } from "lemmy-js-client";
-
-export default function editPostReport(
-  data: PostReportView,
-  reports: PostReportView[],
-) {
-  return editListImmutable("post_report", data, reports);
-}
