--- src/shared/utils/app/enable-downvotes.ts
@@ -1,5 +1,0 @@
-import { GetSiteResponse } from "lemmy-js-client";
-
-export default function enableDownvotes(siteRes: GetSiteResponse): boolean {
-  return siteRes.site_view.local_site.enable_downvotes;
-}
