--- src/shared/utils/app/fetch-theme-list.ts
@@ -1,3 +1,0 @@
-export default async function fetchThemeList(): Promise<string[]> {
-  return fetch("/css/themelist").then(res => res.json());
-}
