--- src/shared/utils/app/fetch-users.ts
@@ -1,7 +1,0 @@
-import { fetchSearchResults } from "@utils/app";
-
-export default async function fetchUsers(q: string) {
-  const res = await fetchSearchResults(q, "Users");
-
-  return res.state === "success" ? res.data.users : [];
-}
