--- src/shared/utils/app/get-comment-id-from-props.ts
@@ -1,8 +1,0 @@
-import { RouteComponentProps } from "inferno-router/dist/Route";
-
-export default function getCommentIdFromProps(
-  props: Pick<RouteComponentProps<{ comment_id?: string }>, "match">,
-): number | undefined {
-  const id = props.match.params.comment_id;
-  return id ? Number(id) : undefined;
-}
