--- src/shared/utils/app/get-data-type-string.ts
@@ -1,5 +1,0 @@
-import { DataType } from "../../interfaces";
-
-export default function getDataTypeString(dt: DataType) {
-  return dt === DataType.Post ? "Post" : "Comment";
-}
