--- src/shared/utils/app/get-id-from-props.ts
@@ -1,8 +1,0 @@
-import { RouteComponentProps } from "inferno-router/dist/Route";
-
-export default function getIdFromProps(
-  props: Pick<RouteComponentProps<{ post_id?: string }>, "match">,
-): number | undefined {
-  const id = props.match.params.post_id;
-  return id ? Number(id) : undefined;
-}
