--- src/shared/utils/app/initialize-site.ts
@@ -1,8 +1,0 @@
-import { GetSiteResponse } from "lemmy-js-client";
-import { setupMarkdown } from "../../markdown";
-import { UserService } from "../../services";
-
-export default function initializeSite(site?: GetSiteResponse) {
-  UserService.Instance.myUserInfo = site?.my_user;
-  setupMarkdown();
-}
