--- src/shared/utils/app/instance-to-choice.ts
@@ -1,9 +1,0 @@
-import { Choice } from "@utils/types";
-import { Instance } from "lemmy-js-client";
-
-export default function instanceToChoice({ id, domain }: Instance): Choice {
-  return {
-    value: id.toString(),
-    label: domain,
-  };
-}
