--- src/shared/utils/app/is-anonymous-path.ts
@@ -1,5 +1,0 @@
-export default function isAnonymousPath(pathname: string) {
-  return /^\/(login.*|signup|password_change.*|verify_email.*)\b/g.test(
-    pathname,
-  );
-}
