--- src/shared/utils/app/is-auth-path.ts
@@ -1,5 +1,0 @@
-export default function isAuthPath(pathname: string) {
-  return /^\/(create_.*?|inbox|settings|admin|reports|registration_applications|activitypub.*?)\b/g.test(
-    pathname,
-  );
-}
