--- src/shared/utils/app/new-vote.ts
@@ -1,9 +1,0 @@
-import { VoteType } from "../../interfaces";
-
-export default function newVote(voteType: VoteType, myVote?: number): number {
-  if (voteType === VoteType.Upvote) {
-    return myVote === 1 ? 0 : 1;
-  } else {
-    return myVote === -1 ? 0 : -1;
-  }
-}
